requirements.txt
python-dotenv
openai
pypdf
python-docx
pandas
tqdm
pillow
pytesseract
tiktoken